feat: extracting dropdown to its own component
